requirements.txt
Django==4.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pillow==9.5.0
djangorestframework==3.14.0
markdown==3.4.3
django-filter==23.2
drf-yasg==1.21.5
djangorestframework-simplejwt==5.2.2